refactor(frontend): extract useYieldChartData hook into its own module
Move the useYieldChartData data-shaping hook (and its YieldChartColumn /
YieldChartCulture types) out of YieldOverview into src/pages/useYieldChartData.ts,
isolating the week/month grouping and legend-ordering logic. The page drops
the parseDateString, formatDateToAPI, formatIsoWeek, mergeCultureYields, and
YieldCalendarCulture imports that were only used by the hook. No behavior
change.
